Refactor: new packages for tool settings and renderers.
