+++ src/main/java/com/esgdev/sparkpaint/engine/tools/FilterBrushTool.java
@@ -2,6 +2,7 @@ package com.esgdev.sparkpaint.engine.tools;
 
 import com.esgdev.sparkpaint.engine.DrawingCanvas;
 import com.esgdev.sparkpaint.engine.selection.Selection;
+import com.esgdev.sparkpaint.engine.tools.renderers.FilterBrushRenderer;
 
 import java.awt.*;
 import java.awt.event.MouseEvent;
