+++ src/main/java/com/esgdev/sparkpaint/engine/tools/settings/BaseToolSettings.java
@@ -1,4 +1,4 @@
-package com.esgdev.sparkpaint.engine.tools;
+package com.esgdev.sparkpaint.engine.tools.settings;
 
 import com.esgdev.sparkpaint.engine.DrawingCanvas;
