+++ src/main/java/com/esgdev/sparkpaint/engine/tools/settings/FreehandSelectionToolSettings.java
@@ -1,6 +1,8 @@
-package com.esgdev.sparkpaint.engine.tools;
+package com.esgdev.sparkpaint.engine.tools.settings;
 
 import com.esgdev.sparkpaint.engine.DrawingCanvas;
+import com.esgdev.sparkpaint.engine.tools.FreeHandSelectionTool;
+import com.esgdev.sparkpaint.engine.tools.ToolManager;
 import com.esgdev.sparkpaint.ui.IconLoader;
 
 import javax.swing.*;
