+++ src/main/java/com/esgdev/sparkpaint/ui/DrawingSettingsToolBox.java
@@ -2,6 +2,7 @@ package com.esgdev.sparkpaint.ui;
 
 import com.esgdev.sparkpaint.engine.DrawingCanvas;
 import com.esgdev.sparkpaint.engine.tools.*;
+import com.esgdev.sparkpaint.engine.tools.settings.*;
 
 import javax.swing.*;
 import java.awt.*;
